improve performance by reusing request buffer allocation
A Vec<u8> allocated is used read the incoming request from the kernel
and stored inside the request. Once the request is serviced this buffer
get dropped.  Allocation can be expensive in a heavily loaded systems,
doing multiple read and writes and may be concurrently.

Another bottleneck: this allocaiton is done on the main thread which
holds the session further hampering response times.

This patch add a new `Mutex`d `VecDeque` of `Vec<u8>` to the `SessionInner`,
Which is shared across the threads through `Arc` thus there is only one
copy of this vector. This VecDeque acts as cache hold pointers the
to already allocated buffers, which are no longer in use.

In order the retain the buffer, this patch add a `Drop` implementation
for the `Request` struct. Once the request is about to be dropped its
assigned buffer is moved to session's VecDeque via `mem::replace`.
There is some sanitization needed for the allocated buffer. This
sanitization is just `clear` its content and using `set_len` to make its
full capacity available for subsequent read.

Thus each time there is a new request we can check with the session's
cache of buffers and if there is none only then allocate.